separate async part in nccl group to avoid circle
