+++ tensorflow/compiler/xla/service/gpu/alpa_nccl_wrapper.h
@@ -24,11 +24,9 @@ limitations under the License.
 #include "third_party/nccl/nccl.h"
 #endif
 
-#include "tensorflow/compiler/xla/pjrt/pjrt_stream_executor_client.h"
 #include "tensorflow/compiler/xla/python/py_buffer.h"
 #include "tensorflow/compiler/xla/python/py_client.h"
-#include "tensorflow/compiler/xla/service/gpu/nccl_utils.h"
-#include "tensorflow/compiler/xla/service/rendezvous.h"
+#include "tensorflow/compiler/xla/service/gpu/alpa_nccl_group_base.h"
 
 namespace xla {
 namespace gpu {
@@ -36,16 +34,9 @@ namespace alpa {
 using AlpaNcclUid = std::vector<int8_t>;
 using AlpaUuids = std::vector<int>;
 
-class CommGroup {
+class PyCommGroup : public CommGroup {
  public:
-  CommGroup(std::shared_ptr<PyClient> backend);
-  // Communicator related functions:
-  Status NcclCreateCommunicators(int world_size,
-                                 const std::vector<int> &device_global_ranks,
-                                 const std::vector<int> &device_ids,
-                                 const AlpaNcclUid &nccl_uid_vec);
-
-  Status NcclDestroyComms(const AlpaNcclUid &storage);
+  PyCommGroup(std::shared_ptr<PyClient> backend);
 
   // Communication operations:
   Status NcclLocalAllGather(const AlpaNcclUid &key,
@@ -77,16 +68,6 @@ class CommGroup {
   void CommWaitCompute(bool is_send, bool is_compute, int device_id);
 
   void ComputeWaitComm(bool is_send, bool is_compute, int device_id);
-
-  NcclComm::Lock AcquireComm(const AlpaNcclUid &uuids, int device_id);
-
- private:
-  std::vector<std::unique_ptr<se::Stream>> send_streams, recv_streams;
-  ThreadSafeMap<std::pair<AlpaNcclUid, int>, NcclComm> comm_map;
-  absl::flat_hash_map<AlpaNcclUid, std::vector<int>> local_ids;
-
-  std::vector<se::StreamExecutor *> executors;
-  PjRtStreamExecutorClient *client;
 };
 
 // We add them here rather than alpa_events to avoid circular deps in Bazel:
@@ -95,20 +76,10 @@ class CommGroup {
 Status ComputationWaitEvents(const AlpaUuids &uuids,
                              std::shared_ptr<PyClient> client);
 
-// Cross-mesh allreduce thunk related
-void SetCommGroup(std::string key, CommGroup *g, const AlpaNcclUid uid);
-
-NcclComm::Lock GetCommunicator(std::string key, size_t device_id);
-
 // Event context management
 void ResetEventContext(std::shared_ptr<PyClient> client);
-// Other functions:
-ncclUniqueId NcclUidDeserialize(const AlpaNcclUid& nccl_uid_chars);
 
-StatusOr<AlpaNcclUid> NcclGetUniqueId();
-
-StatusOr<int> NcclGetVersion();
-
+// Other functions
 StatusOr<int> GetBufferDeviceId(PyBuffer::object buffer);
 }  // namespace alpa
 }  // namespace gpu
